requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
requests
sqlalchemy
google-cloud-secret-manager